+++ changelog.md
@@ -2,6 +2,10 @@
 
 - Bump hahomematic to 2023.7.0
   - Project file maintenance
+- Add services that return data:
+  - get_device_value
+  - get_paramset
+  - get_variable_value
 - Fix entity state translations
 
 # Version 1.38.0 (2023-07-02)